Multinegocio Fases 0-3: settings por negocio, impuestos flexibles, catálogo decimal, verticales farmacia/celulares/EAN
- Fase 0: SettingsRepository/Service/Controller + SettingsPage, ticket y
  reportes con cabecera del negocio, Money con simbolo/decimales.
- Fase 1: resolveTaxRate contra tax_rates_json, desglose por tasa
  (Totals/buckets, sales.tax_breakdown, ticket, POS, reportes).
- Fase 2: categories + CategoryRepository, qty/stock double (granel),
  unidades validadas, filtro POS, 14 seeds por rubro opt-in.
- Fase 3: attrs_json, lote/vencimiento obligatorio, bloqueo vencidos,
  receta + controlado con supervisor, seriales/IMEI + garantia/RMA,
  EAN-13. Suites tst_settings/tst_tax/tst_catalog/tst_vertical en verde.

+++ src/controllers/OpsControllers.cpp
@@ -33,7 +33,7 @@ void InventoryController::refresh()
     emit movementsChanged();
 }
 
-QVariantMap InventoryController::adjust(const QString &sku, int delta, const QString &reason,
+QVariantMap InventoryController::adjust(const QString &sku, double delta, const QString &reason,
                                         const QString &user)
 {
     const auto r = m_service->registerAdjustment(sku, delta, reason, user);
@@ -43,7 +43,7 @@ QVariantMap InventoryController::adjust(const QString &sku, int delta, const QSt
     return {{"ok", true}, {"newStock", r.value().newStock}};
 }
 
-QVariantMap InventoryController::transfer(const QString &sku, int qty, const QString &to,
+QVariantMap InventoryController::transfer(const QString &sku, double qty, const QString &to,
                                           const QString &reason, const QString &user)
 {
     const auto r = m_service->transfer(sku, qty, to, reason, user);
@@ -86,7 +86,7 @@ void PurchasesController::refresh()
     emit ordersChanged();
 }
 
-QVariantMap PurchasesController::create(const QString &supplier, const QString &sku, int qty,
+QVariantMap PurchasesController::create(const QString &supplier, const QString &sku, double qty,
                                         const QString &user)
 {
     const auto r = m_service->create(supplier, sku, qty, user);
